Move all DB-related code from SongLoader to new classes, provider.StoreLoader and discog.MediaStoreBridge
